Add FileOperations to web
